requirements.txt
gspread==5.7.2
google-auth==2.21.0
